Switch back to panics for enum exhaustiveness
For a bunch of reasons I generally approve of, Go doesn't provide
Rust- or Swift-style exhaustiveness checks for switches on enums. I
tried using the Antithesis SDK's Unreachable assertions in some default
branches, but (1) it's difficult to make the text intelligible in
reports, and (2) assertions don't actually exit/panic/return, so we
still need some behavior defined.

On balance, it seems better to either panic or return errors. This code
is all testing helpers, so panics are simpler - and in a demo project,
I'm willing to keep them.